# perf: Optimize shuffle cache